+++ bridge/core/dom/node_test.cc
@@ -15,7 +15,7 @@ TEST(Node, appendChild) {
     EXPECT_STREQ(message.c_str(), "true true true");
     logCalled = true;
   };
-  auto env = TEST_init([](int32_t contextId, const char* errmsg) { errorCalled = true; });
+  auto env = TEST_init([](double contextId, const char* errmsg) { errorCalled = true; });
   auto context = env->page()->GetExecutingContext();
   const char* code =
       "let div = document.createElement('div');"
@@ -81,7 +81,7 @@ TEST(Node, nodeName) {
     EXPECT_STREQ(message.c_str(), "DIV #text #document-fragment #comment #document");
     logCalled = true;
   };
-  auto env = TEST_init([](int32_t contextId, const char* errmsg) { errorCalled = true; });
+  auto env = TEST_init([](double contextId, const char* errmsg) { errorCalled = true; });
   auto context = env->page()->GetExecutingContext();
   const char* code =
       "let div = document.createElement('div');"
@@ -102,7 +102,7 @@ TEST(Node, childNodes) {
     EXPECT_STREQ(message.c_str(), "true true true true");
     logCalled = true;
   };
-  auto env = TEST_init([](int32_t contextId, const char* errmsg) { errorCalled = true; });
+  auto env = TEST_init([](double contextId, const char* errmsg) { errorCalled = true; });
   auto context = env->page()->GetExecutingContext();
   MemberMutationScope scope{context};
   const char* code =
@@ -125,7 +125,7 @@ TEST(Node, textNodeHaveEmptyChildNodes) {
   bool static errorCalled = false;
   bool static logCalled = false;
   webf::WebFPage::consoleMessageHandler = [](void* ctx, const std::string& message, int logLevel) { logCalled = true; };
-  auto env = TEST_init([](int32_t contextId, const char* errmsg) { errorCalled = true; });
+  auto env = TEST_init([](double contextId, const char* errmsg) { errorCalled = true; });
   auto context = env->page()->GetExecutingContext();
   const char* code =
       "let text = document.createTextNode('helloworld');"
@@ -143,7 +143,7 @@ TEST(Node, textContent) {
     EXPECT_STREQ(message.c_str(), "1234helloworld");
     logCalled = true;
   };
-  auto env = TEST_init([](int32_t contextId, const char* errmsg) { errorCalled = true; });
+  auto env = TEST_init([](double contextId, const char* errmsg) { errorCalled = true; });
   auto context = env->page()->GetExecutingContext();
   const char* code =
       "let text1 = document.createTextNode('1234');"
@@ -165,7 +165,7 @@ TEST(Node, setTextContent) {
     EXPECT_STREQ(message.c_str(), "1234");
     logCalled = true;
   };
-  auto env = TEST_init([](int32_t contextId, const char* errmsg) { errorCalled = true; });
+  auto env = TEST_init([](double contextId, const char* errmsg) { errorCalled = true; });
   auto context = env->page()->GetExecutingContext();
   const char* code =
       "let div = document.createElement('div');"
@@ -184,7 +184,7 @@ TEST(Node, ensureDetached) {
     EXPECT_STREQ(message.c_str(), "true true");
     logCalled = true;
   };
-  auto env = TEST_init([](int32_t contextId, const char* errmsg) { errorCalled = true; });
+  auto env = TEST_init([](double contextId, const char* errmsg) { errorCalled = true; });
   auto context = env->page()->GetExecutingContext();
   const char* code =
       "let div = document.createElement('div');"
@@ -203,7 +203,7 @@ TEST(Node, replaceBody) {
   bool static errorCalled = false;
   bool static logCalled = false;
   webf::WebFPage::consoleMessageHandler = [](void* ctx, const std::string& message, int logLevel) { logCalled = true; };
-  auto env = TEST_init([](int32_t contextId, const char* errmsg) {
+  auto env = TEST_init([](double contextId, const char* errmsg) {
     WEBF_LOG(VERBOSE) << errmsg;
     errorCalled = true;
   });
@@ -241,7 +241,7 @@ TEST(Node, cloneNode) {
     logCalled = true;
     EXPECT_STREQ(message.c_str(), "true true true");
   };
-  auto env = TEST_init([](int32_t contextId, const char* errmsg) {
+  auto env = TEST_init([](double contextId, const char* errmsg) {
     WEBF_LOG(VERBOSE) << errmsg;
     errorCalled = true;
   });
@@ -290,7 +290,7 @@ TEST(Node, nestedNode) {
     logCalled = true;
     EXPECT_STREQ(message.c_str(), "true true true");
   };
-  auto env = TEST_init([](int32_t contextId, const char* errmsg) {
+  auto env = TEST_init([](double contextId, const char* errmsg) {
     WEBF_LOG(VERBOSE) << errmsg;
     errorCalled = true;
   });
@@ -332,7 +332,7 @@ el.replaceChild(child_3, child_1);
     logCalled = true;
     EXPECT_STREQ(message.c_str(), "true true true");
   };
-  auto env = TEST_init([](int32_t contextId, const char* errmsg) {
+  auto env = TEST_init([](double contextId, const char* errmsg) {
     WEBF_LOG(VERBOSE) << errmsg;
     errorCalled = true;
   });
@@ -358,7 +358,7 @@ console.assert(el.isConnected == false);
     logCalled = true;
     EXPECT_STREQ(message.c_str(), "true true true");
   };
-  auto env = TEST_init([](int32_t contextId, const char* errmsg) {
+  auto env = TEST_init([](double contextId, const char* errmsg) {
     WEBF_LOG(VERBOSE) << errmsg;
     errorCalled = true;
   });
