+++ bridge/core/html/custom/widget_element_test.cc
@@ -15,7 +15,7 @@ TEST(WidgetElement, setPropertyEventHandler) {
     EXPECT_STREQ(message.c_str(), "1111");
     logCalled = true;
   };
-  auto env = TEST_init([](int32_t contextId, const char* errmsg) {
+  auto env = TEST_init([](double contextId, const char* errmsg) {
     WEBF_LOG(VERBOSE) << errmsg;
     errorCalled = true;
   });
@@ -37,7 +37,7 @@ TEST(WidgetElement, getPropertyWithSymbolToStringTag) {
     EXPECT_STREQ(message.c_str(), "FLUTTER-CHECKBOX");
     logCalled = true;
   };
-  auto env = TEST_init([](int32_t contextId, const char* errmsg) {
+  auto env = TEST_init([](double contextId, const char* errmsg) {
     WEBF_LOG(VERBOSE) << errmsg;
     errorCalled = true;
   });
