+++ bridge/core/timing/performance_test.cc
@@ -15,7 +15,7 @@ TEST(Performance, now) {
     logCalled = true;
     EXPECT_STREQ(message.c_str(), "true");
   };
-  auto env = TEST_init([](int32_t contextId, const char* errmsg) {
+  auto env = TEST_init([](double contextId, const char* errmsg) {
     WEBF_LOG(VERBOSE) << errmsg;
     errorCalled = true;
   });
@@ -33,7 +33,7 @@ TEST(Performance, timeOrigin) {
     logCalled = true;
     EXPECT_STREQ(message.c_str(), "true");
   };
-  auto env = TEST_init([](int32_t contextId, const char* errmsg) {
+  auto env = TEST_init([](double contextId, const char* errmsg) {
     WEBF_LOG(VERBOSE) << errmsg;
     errorCalled = true;
   });
@@ -51,7 +51,7 @@ TEST(Performance, toJSON) {
     logCalled = true;
     EXPECT_STREQ(message.c_str(), "true true");
   };
-  auto env = TEST_init([](int32_t contextId, const char* errmsg) {
+  auto env = TEST_init([](double contextId, const char* errmsg) {
     WEBF_LOG(VERBOSE) << errmsg;
     errorCalled = true;
   });
@@ -71,7 +71,7 @@ TEST(Performance, mark) {
     logCalled = true;
     EXPECT_STREQ(message.c_str(), "PerformanceMark {detail: null}");
   };
-  auto env = TEST_init([](int32_t contextId, const char* errmsg) {
+  auto env = TEST_init([](double contextId, const char* errmsg) {
     WEBF_LOG(VERBOSE) << errmsg;
     errorCalled = true;
   });
@@ -92,7 +92,7 @@ TEST(Performance, markWithDetail) {
     logCalled = true;
     EXPECT_STREQ(message.c_str(), "PerformanceMark {detail: {...}}");
   };
-  auto env = TEST_init([](int32_t contextId, const char* errmsg) {
+  auto env = TEST_init([](double contextId, const char* errmsg) {
     WEBF_LOG(VERBOSE) << errmsg;
     errorCalled = true;
   });
@@ -113,7 +113,7 @@ TEST(Performance, markWithName) {
     logCalled = true;
     EXPECT_STREQ(message.c_str(), "[PerformanceMark {...}, PerformanceMark {...}]");
   };
-  auto env = TEST_init([](int32_t contextId, const char* errmsg) {
+  auto env = TEST_init([](double contextId, const char* errmsg) {
     WEBF_LOG(VERBOSE) << errmsg;
     errorCalled = true;
   });
@@ -136,7 +136,7 @@ TEST(Performance, clearMarks) {
     logCalled = true;
     EXPECT_STREQ(message.c_str(), "[]");
   };
-  auto env = TEST_init([](int32_t contextId, const char* errmsg) {
+  auto env = TEST_init([](double contextId, const char* errmsg) {
     WEBF_LOG(VERBOSE) << errmsg;
     errorCalled = true;
   });
@@ -160,7 +160,7 @@ TEST(Performance, clearMarksByName) {
     logCalled = true;
     EXPECT_STREQ(message.c_str(), "[PerformanceMark {...}]");
   };
-  auto env = TEST_init([](int32_t contextId, const char* errmsg) {
+  auto env = TEST_init([](double contextId, const char* errmsg) {
     WEBF_LOG(VERBOSE) << errmsg;
     errorCalled = true;
   });
@@ -184,7 +184,7 @@ TEST(Performance, measure) {
     logCalled = true;
     EXPECT_STREQ(message.c_str(), "true");
   };
-  auto env = TEST_init([](int32_t contextId, const char* errmsg) {
+  auto env = TEST_init([](double contextId, const char* errmsg) {
     WEBF_LOG(VERBOSE) << errmsg;
     errorCalled = true;
   });
